Refactor Planet structures to new HATEOAS architecture.
