+++ api/src/main/java/com/senorpez/trident/api/EmptyAPIResources.java
@@ -1,14 +1,14 @@
 package com.senorpez.trident.api;
 
+import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.Link;
-import org.springframework.hateoas.Resources;
 
-import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
-import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
-class EmptyAPIResources extends Resources<Object> {
+class EmptyAPIResources extends CollectionModel<Object> {
     EmptyAPIResources(Iterable<Object> content, Link... links) {
-        super(content, links);
+        CollectionModel.of(content, links);
         this.add(linkTo(methodOn(RootController.class).root()).withRel("index"));
     }
 }
