+++ api/src/main/java/com/senorpez/trident/api/EmptyPlanetResources.java
@@ -2,8 +2,8 @@ package com.senorpez.trident.api;
 
 import org.springframework.hateoas.Link;
 
-import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
-import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 class EmptyPlanetResources extends EmptyAPIResources {
     EmptyPlanetResources(Iterable<Object> content, int solarSystemId, int starId, Link... links) {
