+++ api/src/main/java/com/senorpez/trident/api/PlanetController.java
@@ -1,10 +1,10 @@
 package com.senorpez.trident.api;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.hateoas.Resource;
-import org.springframework.hateoas.Resources;
-import org.springframework.hateoas.core.EmbeddedWrapper;
-import org.springframework.hateoas.core.EmbeddedWrappers;
+import org.springframework.hateoas.CollectionModel;
+import org.springframework.hateoas.RepresentationModel;
+import org.springframework.hateoas.server.core.EmbeddedWrapper;
+import org.springframework.hateoas.server.core.EmbeddedWrappers;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,12 +16,12 @@ import java.util.Collections;
 import java.util.stream.Collectors;
 
 import static com.senorpez.trident.api.SupportedMediaTypes.TRIDENT_API_VALUE;
-import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 @RequestMapping(
         value = "/systems/{solarSystemId}/stars/{starId}/planets",
         method = RequestMethod.GET,
-        produces = {TRIDENT_API_VALUE, APPLICATION_JSON_UTF8_VALUE}
+        produces = {TRIDENT_API_VALUE, APPLICATION_JSON_VALUE}
 )
 @RestController
 class PlanetController {
@@ -35,7 +35,7 @@ class PlanetController {
     }
 
     @RequestMapping
-    ResponseEntity<? extends Resources> planets(@PathVariable final int solarSystemId, @PathVariable final int starId) {
+    ResponseEntity<? extends RepresentationModel<?>> planets(@PathVariable final int solarSystemId, @PathVariable final int starId) {
         final SolarSystem solarSystem = apiService.findOne(
                 this.solarSystems,
                 findSolarSystem -> findSolarSystem.getId() == solarSystemId,
@@ -46,22 +46,22 @@ class PlanetController {
                 () -> new StarNotFoundException(starId));
         if (star.getPlanets() == null) {
             EmbeddedWrappers wrappers = new EmbeddedWrappers(false);
-            EmbeddedWrapper wrapper = wrappers.emptyCollectionOf(PlanetModel.class);
-            return ResponseEntity.ok(new EmptyPlanetResources(Collections.singletonList(wrapper), solarSystemId, starId));
+            EmbeddedWrapper wrapper = wrappers.emptyCollectionOf(PlanetEntity.class);
+            return ResponseEntity.ok(CollectionModel.of(new EmptyPlanetResources(Collections.singletonList(wrapper), solarSystemId, starId)));
         } else {
             final Collection<Planet> planets = star.getPlanets();
-            final Collection<EmbeddedPlanetModel> planetModels = planets.stream()
-                    .map(EmbeddedPlanetModel::new)
-                    .collect(Collectors.toList());
-            final Collection<Resource<EmbeddedPlanetModel>> planetResources = planetModels.stream()
-                    .map(planetModel -> planetModel.toResource(solarSystemId, starId))
-                    .collect(Collectors.toList());
-            return ResponseEntity.ok(new EmbeddedPlanetResources(planetResources, solarSystemId, starId));
+            final CollectionModel<PlanetModel> planetModels = CollectionModel.of(planets
+                    .stream()
+                    .map(EmbeddedPlanetEntity::new)
+                    .map(embeddedPlanetEntity -> new PlanetModel(embeddedPlanetEntity, solarSystemId, starId))
+                    .collect(Collectors.toList())
+            );
+            return ResponseEntity.ok(planetModels);
         }
     }
 
     @RequestMapping("/{planetId}")
-    ResponseEntity<PlanetResource> planets(@PathVariable final int solarSystemId, @PathVariable final int starId, @PathVariable final int planetId) {
+    ResponseEntity<PlanetModel> planets(@PathVariable final int solarSystemId, @PathVariable final int starId, @PathVariable final int planetId) {
         final SolarSystem solarSystem = apiService.findOne(
                 this.solarSystems,
                 findSolarSystem -> findSolarSystem.getId() == solarSystemId,
@@ -74,8 +74,8 @@ class PlanetController {
                 star.getPlanets(),
                 findPlanet -> findPlanet.getId() == planetId,
                 () -> new PlanetNotFoundException(planetId));
-        final PlanetModel planetModel = new PlanetModel(planet);
-        final PlanetResource planetResource = planetModel.toResource(solarSystemId, starId);
-        return ResponseEntity.ok(planetResource);
+        final PlanetEntity planetEntity = new PlanetEntity(planet);
+        final PlanetModel planetModel = new PlanetModel(planetEntity, solarSystemId, starId);
+        return ResponseEntity.ok(planetModel);
     }
 }
